refactor(api): [#143] extract api mods
